# docs/api/openapi/processor.json
{"openapi":"3.1.0","info":{"title":"Spotify Memory System — Memory Processor","description":"Extraction, entity resolution, policy gate, and temporal graph writes.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/memories/extract":{"post":{"summary":"Extract","description":"Deterministic validation entry point — converts one approved event into\ntyped candidates WITHOUT writing to the graph (§7.3 endpoint #2).","operationId":"extract_v1_memories_extract_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InteractionEvent"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExtractionResult"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/memories":{"post":{"summary":"Create Memory","description":"Create an explicit or approved memory (§7.3 endpoint #3). This is the\nONLY path that writes a new durable fact — never write solely because an\nLLM generated it in a response (§7.5 Memory safety).","operationId":"create_memory_v1_memories_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoryCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/memories/{memory_id}":{"patch":{"summary":"Correct Memory","description":"Correct, expire, or reactivate a memory under optimistic concurrency\n(§7.3 endpoint #6). Corrections SUPERSEDE — they never overwrite history.","operationId":"correct_memory_v1_memories__memory_id__patch","parameters":[{"name":"memory_id","in":"path","required":true,"schema":{"type":"string","title":"Memory Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoryCorrectionRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ExtractionCandidate":{"properties":{"memory_id":{"type":"string","title":"Memory Id","description":"Proposed stable id (deterministic hash of subject+fact+source)"},"decision":{"type":"string","enum":["accept","reject","needs_confirmation"],"title":"Decision"},"normalized_fact":{"type":"string","title":"Normalized Fact"},"entities":{"items":{"type":"string"},"type":"array","title":"Entities"},"memory_type":{"type":"string","enum":["episode","explicit_preference","candidate_preference","exclusion","correction"],"title":"Memory Type"},"relevance_score":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Relevance Score"},"confidence":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence"},"temporal_scope":{"type":"string","enum":["episodic","durable"],"title":"Temporal Scope"},"policy_flags":{"items":{"type":"string"},"type":"array","title":"Policy Flags","description":"e.g. ['sensitive_inference_blocked']"},"policy_class":{"type":"string","enum":["normal","sensitive","blocked"],"title":"Policy Class","default":"normal"},"reason":{"type":"string","title":"Reason","description":"Why this decision was made — required for provenance"},"source_event_id":{"type":"string","title":"Source Event Id"}},"type":"object","required":["memory_id","decision","normalized_fact","memory_type","relevance_score","confidence","temporal_scope","reason","source_event_id"],"title":"ExtractionCandidate","description":"One candidate memory proposed from a single interaction event.\nFields mirror §7.5's required structured-output fields exactly:\nmemory_id, decision, normalized_fact, entities, relevance_score, confidence,\ntemporal_scope, policy_flags, reason."},"ExtractionResult":{"properties":{"event_id":{"type":"string","title":"Event Id"},"candidates":{"items":{"$ref":"#/components/schemas/ExtractionCandidate"},"type":"array","title":"Candidates"},"rejected_reason":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Rejected Reason","description":"Set when the event yields zero candidates"}},"type":"object","required":["event_id"],"title":"ExtractionResult"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InteractionEvent":{"properties":{"schema_version":{"type":"string","title":"Schema Version","description":"Contract version for compatibility checks","default":"1.0.0"},"event_id":{"type":"string","title":"Event Id","description":"Unique source-event identifier"},"subject_id":{"type":"string","title":"Subject Id","description":"Authenticated subject (user) scope"},"surface":{"type":"string","title":"Surface","description":"Originating AI surface, e.g. 'music_chat', 'playlist_ui', 'podcast_chat'"},"event_type":{"type":"string","enum":["play","save","follow","skip","statement","correction","opt_out","delete_request"],"title":"Event Type","description":"Interaction classification"},"payload":{"additionalProperties":true,"type":"object","title":"Payload","description":"Event-type-specific structured payload"},"locale":{"type":"string","title":"Locale","description":"Locale context of the interaction","default":"en-US"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"Client-reported time of occurrence (UTC)"},"consent_state":{"type":"string","enum":["granted","denied","partial"],"title":"Consent State","description":"Consent status at time of capture"},"idempotency_key":{"type":"string","title":"Idempotency Key","description":"Client-supplied key; duplicate keys must not double-write"}},"type":"object","required":["event_id","subject_id","surface","event_type","timestamp","consent_state","idempotency_key"],"title":"InteractionEvent","description":"A single eligible interaction event entering the system via the Ingestion API.\n\nEvery field here is required by §5.4: subject scope, surface, locale, timestamp,\nconsent state, source-event identifier, and idempotency key must be attached to\nevery event before it is allowed into graph processing."},"MemoryCorrectionRequest":{"properties":{"action":{"type":"string","enum":["correct","expire","reactivate"],"title":"Action"},"new_fact_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"New Fact Text","description":"Required when action == 'correct'"},"reason":{"type":"string","title":"Reason","description":"Human-readable justification, stored for audit"},"expected_version":{"type":"string","title":"Expected Version","description":"Optimistic-concurrency token from the last read"}},"type":"object","required":["action","reason","expected_version"],"title":"MemoryCorrectionRequest","description":"Input for PATCH /v1/memories/{memory_id} — correct, supersede, or expire."},"MemoryCreateRequest":{"properties":{"subject_id":{"type":"string","title":"Subject Id"},"fact_text":{"type":"string","title":"Fact Text"},"memory_type":{"type":"string","enum":["episode","explicit_preference","candidate_preference","exclusion","correction"],"title":"Memory Type"},"entities":{"items":{"type":"string"},"type":"array","title":"Entities"},"confidence":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Confidence"},"source_event_id":{"type":"string","title":"Source Event Id"},"surface":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Surface"}},"type":"object","required":["subject_id","fact_text","memory_type","confidence","source_event_id"],"title":"MemoryCreateRequest","description":"Input for POST /v1/memories — create an explicit or approved memory."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}